vibe-compass-bot/requirements.txt
# Основные зависимости для Telegram бота (обновлены для совместимости с Python 3.14)
python-telegram-bot>=21.0
python-dotenv==1.0.0

# База данных (обновлена для совместимости с Python 3.14)
sqlalchemy>=2.0.35

# HTTP запросы для CRM интеграции
requests==2.31.0

# Дополнительные утилиты
python-dateutil==2.8.2
